myhub-tui/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "myhub-tui"
version = "0.1.0"
description = "Terminal UI for myhub — the SSD-portable personal AI hub. Ports the OpenAra UX pattern (prompt_toolkit + rich, pending-handler wizards, command registry) to a macOS-native context."
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Kolja Schoepe" }]
classifiers = [
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Terminals",
]
dependencies = [
    "prompt-toolkit>=3.0.0",
    "rich>=13.0.0",
    "psutil>=5.9.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.8.0"]

[project.scripts]
myhub-tui = "myhub_tui.app:run"

[tool.setuptools.packages.find]
include = ["myhub_tui*"]

[tool.ruff]
target-version = "py313"
line-length = 100
